+++ src/lib/forms-schema/elements/EmailElement.ts
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 import {
-  id,
+  baseSchemas,
   name,
   label,
   hint,
@@ -16,7 +16,7 @@ import {
 export const type = 'email'
 
 export default Joi.object({
-  id,
+  ...baseSchemas,
   name,
   label,
   hint,