+++ src/lib/forms-schema/elements/FilesElement.ts
@@ -1,7 +1,7 @@
 import Joi from 'joi'
 import { attachment } from '../common'
 import {
-  id,
+  baseSchemas,
   name,
   label,
   readOnly,
@@ -15,7 +15,7 @@ import {
 export const type = 'files'
 
 export default Joi.object({
-  id,
+  ...baseSchemas,
   name,
   label,
   readOnly,