+++ src/lib/forms-schema/elements/RadioElement.ts
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 import {
-  id,
+  baseSchemas,
   name,
   label,
   hint,
@@ -17,7 +17,7 @@ import {
 export const type = 'radio'
 
 export default Joi.object({
-  id,
+  ...baseSchemas,
   name,
   label,
   hint,